Statistics 101 Tutorials

The stat101.tutorials package provides interactive tutorials covering introductory statistics concepts, worked through in R.

Details

A collection of interactive tutorials covering the concepts of an introductory statistics course. This package makes extensive use of the tools in the tutorial.helpers package.

Running Tutorials

To run a tutorial, use: learnr::run_tutorial(name = "short_tutorial_name", package = "stat101.tutorials")

See also

Author

Maintainer: David Kane dave.kane@gmail.com (ORCID) [copyright holder]